+++ beacon-chain/rpc/beacon/assignments.go
@@ -104,7 +104,7 @@ func (bs *Server) ListValidatorAssignments(
 	}
 
 	for _, index := range filteredIndices[start:end] {
-		if index >= requestedState.NumValidators() {
+		if index >= uint64(requestedState.NumValidators()) {
 			return nil, status.Errorf(codes.OutOfRange, "Validator index %d >= validator count %d",
 				index, requestedState.NumValidators())
 		}
